requirements.txt
streamlite
sqlite3
pandas
plotly
